% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Takes the output from \code{\link{eq_location_clean}}, expects TSUNAMI, VOLCANO, YEAR, MONTH, DAY, HOUR, MINUTE, and SECOND columns.}
\usage{
eq_clean_data(rawdata = NULL)
}
\arguments{
\item{rawdata}{optional, dataframe outputted from \code{\link{eq_read_data}} or \code{\link{eq_location_clean}}}
}
\description{
Takes the output from \code{\link{eq_location_clean}}, expects TSUNAMI, VOLCANO, YEAR, MONTH, DAY, HOUR, MINUTE, and SECOND columns.
}
\examples{

\dontrun{

# Clean data and filter out entries with no TOTAL_DEATHS data
data <- eq_clean_data() \%>\%
    dplyr::filter(!is.na(TOTAL_DEATHS)) \%>\%
    dplyr::select(DATE, YEAR, COUNTRY, REGION, LONGITUDE, LATITUDE, MAG, TOTAL_DEATHS) \%>\%
    tidyr::drop_na()
}


}
