% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{This function creates HTML to be displayed in the pops generated by the \code{\link{eq_map}} function.}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{dataframe with \code{LOCATION}, \code{MAG}, and \code{TOTAL_DEATHS} columns}
}
\value{
vector containing HTML
}
\description{
This function creates HTML to be displayed in the pops generated by the \code{\link{eq_map}} function.
}
\examples{

\dontrun{
eq_clean_data() \%>\%
      dplyr::filter(!is.na(TOTAL_DEATHS)) \%>\%
      dplyr::select(DATE, YEAR, COUNTRY, REGION, LONGITUDE, LATITUDE, MAG, TOTAL_DEATHS) \%>\%
      tidyr::drop_na() \%>\%
      dplyr::filter(COUNTRY == "MEXICO" & YEAR >= 2000)  \%>\%
      dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
      eq_map(annot_col = "popup_text")
}

}
