% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Formats the LOCATION_NAME column correctly.}
\usage{
eq_location_clean()
}
\value{
dataframe
}
\description{
Formats the LOCATION_NAME column correctly.
}
