% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{eq_read_data}
\alias{eq_read_data}
\title{Reads in the raw data from the `inst/extdata` directory.}
\usage{
eq_read_data()
}
\value{
dataframe
}
\description{
Reads in the raw data from the `inst/extdata` directory.
}
