% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans and formats the rawdata}
\usage{
eq_clean_data(rawdata)
}
\arguments{
\item{rawdata}{dataframe returned from `eq_read_data` function}
}
\description{
Cleans and formats the rawdata
}
