% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Formats the LOCATION column correctly, removing redundant COUNTRY/REGION information.}
\usage{
eq_location_clean(data = NULL)
}
\arguments{
\item{data}{dataframe object}
}
\value{
dataframe
}
\description{
Formats the LOCATION column correctly, removing redundant COUNTRY/REGION information.
}
