% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Takes the output from \code{\link{eq_read_data}}, but will work on any data from with LOCATION column. Creates COUNTRY and REGION columns by splitting input from the LOCATION column which should contain strings of the form \code{"COUNTRY: REGION"} (e.g. \code{"JAPAN:TOKYO"}).}
\usage{
eq_location_clean(data = NULL)
}
\arguments{
\item{data}{a dataframe or tibble with a LOCATION column}
}
\value{
dataframe
}
\description{
Takes the output from \code{\link{eq_read_data}}, but will work on any data from with LOCATION column. Creates COUNTRY and REGION columns by splitting input from the LOCATION column which should contain strings of the form \code{"COUNTRY: REGION"} (e.g. \code{"JAPAN:TOKYO"}).
}
\examples{

\dontrun{

# Retrieve all earthquake data for Japan since 2000
data <- eq_read_data() \%>\%
           filter(COUNTRY == "JAPAN", YEAR >= 2000) \%>\%
           dplyr::filter(!is.na(TOTAL_DEATHS)) \%>\%
           dplyr::select(DATE, YEAR, COUNTRY, REGION, LONGITUDE, LATITUDE, MAG, TOTAL_DEATHS) \%>\%
           tidyr::drop_na()
}

}
