% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_read_data}
\alias{eq_read_data}
\title{Reads in the raw data from the `inst/extdata` directory. Renames some columns, drops some irrelevant columns and converts all column names to uppercase.}
\usage{
eq_read_data(
  filename = system.file("extdata", "earthquakes-2021-03-08_17-02-58_-0500.tsv",
    package = "earthquakes")
)
}
\arguments{
\item{filename}{path to datafile}
}
\value{
dataframe
}
\description{
Reads in the raw data from the `inst/extdata` directory. Renames some columns, drops some irrelevant columns and converts all column names to uppercase.
}
\examples{

\dontrun{

# Retrieve all earthquake data from 2020
data <- eq_read_data() \%>\%
           dplyr::filter(YEAR == 2020)
}

}
