% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plot.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Defines the GeomTimeline ggplot2 geom. It is recommended to use the geom_timeline function to create these layers.}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimeline
}
\value{
Geom
}
\description{
Defines the GeomTimeline ggplot2 geom. It is recommended to use the geom_timeline function to create these layers.
}
\examples{

\dontrun{
# All deadly earthquakes in JAPAN, CHINA and NEPAL since 2000

data <- eq_clean_data() \%>\%
     dplyr::filter(!is.na(TOTAL_DEATHS)) \%>\%
     dplyr::select(DATE, YEAR, COUNTRY, REGION, LONGITUDE, LATITUDE, MAG, TOTAL_DEATHS) \%>\%
     tidyr::drop_na() \%>\%
     dplyr::filter(YEAR > 1999, COUNTRY \%in\% c("JAPAN", "CHINA","NEPAL"))

# Plot Timeline

ggplot2::ggplot() +
    ggplot2::layer(geom = GeomTimeline,
                   mapping = aes(x = data$DATE,
                                 y = data$COUNTRY,
                                 size = data$MAG,
                                 color = data$TOTAL_DEATHS),
                   data = data,
                   stat = 'identity',
                   position = 'identity',
                   show.legend = NA,
                   inherit.aes = TRUE,
                   params = list(na.rm = FALSE))
 }

}
\keyword{datasets}
