% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plot.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{This function creates a vertical lines and labels, for use with the GeomTimeline layer.}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimelineLabel
}
\description{
This function creates a vertical lines and labels, for use with the GeomTimeline layer.
}
\examples{

\dontrun{

# All deadly earthquakes in JAPAN, CHINA and NEPAL since 2000
data <- eq_clean_data() \%>\%
     dplyr::filter(!is.na(TOTAL_DEATHS)) \%>\%
     dplyr::select(DATE, YEAR, COUNTRY, REGION, LONGITUDE, LATITUDE, MAG, TOTAL_DEATHS) \%>\%
     tidyr::drop_na()

    ggplot2::ggplot() +

    ggplot2::layer(geom = GeomTimelineLabel,
                   mapping = aes(x = data$DATE,
                                 y = data$COUNTRY,
                                 size = data$MAG,
                                 color = data$TOTAL_DEATHS),
                   data = data,
                   stat = 'identity',
                   position = 'identity',
                   show.legend = NA,
                   inherit.aes = TRUE,
                   params = list(na.rm = FALSE))
}

}
\keyword{datasets}
